add JacobiSymbol enum

+++ tests/int.rs
@@ -45,8 +45,8 @@ proptest! {
         let g_bi = to_bigint(&g);
 
         let expected = g_bi.jacobi(&f_bi);
-        let actual = g.jacobi_symbol(&f);
-        let actual_vartime = g.jacobi_symbol_vartime(&f);
+        let actual = g.jacobi_symbol(&f) as i8;
+        let actual_vartime = g.jacobi_symbol_vartime(&f) as i8;
         prop_assert_eq!(expected, actual);
         prop_assert_eq!(expected, actual_vartime);
     }
